+++ packages/apollo-client/CHANGELOG.md
@@ -2,6 +2,10 @@
 # Change log
 
 ### vNEXT
+- Add new fetchPolicy called 'no-cache' to bypass reading from or saving to the cache when making a query
+
+### 2.2.1
+- Allow optional parameter to include queries in standby mode when refetching observed queries [PR#2804](https://github.com/apollographql/apollo-client/pull/2804)
 
 ### 2.2.0
 - include `optimisticResponse` in the context passed to apollo-link for mutations [PR#2704](https://github.com/apollographql/apollo-client/pull/2704)